MDEV-19158: MariaDB 10.2.22 is writing duplicate entries into binary log
Problem:
========
We have a Master/Master Setup on two servers, but are only writing to one of
those servers (so it is essentially Master/Slave) We upgraded from 10.1.* to
10.2.22 last week and starting with the upgrade, we are getting duplicate key
errors on the slave.  BINLOG=mixed.

Analysis:
=========
This issue happens with LOCK TABLES and binlog_format=MIXED combination.
When an UNSAFE statement is encountered it is logged in the form of ROW format.
For all the tables that are part of LOCK TABLES list their table maps are
written into the binary log. For each table present in the list a check is done
to see if 'check_table_binlog_row_based_done' flag is set. If it is not set a check
process is initiated to see if table qualifies for row based binary logging or
not and 'check_table_binlog_row_based_done' is set. This flag will be cleared at
the time of closing thread tables.

But there can be special cases where the LOCK TABLES contains more number of
tables but the unsafe query is actually using subset of tables from LOCK TABLES
list.

For example: LOCK TABLES locks t1,t2,t3 but the unsafe statement makes use of
only two tables t1,t3. In this case the 'check_table_binlog_row_based_done' flag
is enabled for table 't2' but 'close_thread_tables' function call will not reset
this flag. This affects the other normal SAFE statements to be forcefully
treated as 'ROW' based statements but they are 'STATEMENT' based. This leads to
an assert on debug builds and causes duplicate entries in release builds. In
release builds a statement is logged in the form of both ROW and STATEMENT
format. This causes the slave to fail with duplicate key error.

Fix:
===
As part of 'close_thread_tables' check if LOCK TABLES is being used. "ha_reset"
is invoked only for those tables which were part of query execution the rest are
left without doing a "ha_reset". As part of fix add an additional check for
"check_table_binlog_row_based_done" flag. If the flag is set invoke "ha_reset"
for that table.